refactor: add folders for canvas and graph classes
