+++ src/main/kotlin/view/MainView.kt
@@ -6,13 +6,10 @@ import androidx.compose.animation.core.animateOffsetAsState
 import androidx.compose.animation.core.tween
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.PointerMatcher
-import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.gestures.detectDragGestures
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.Text
 import androidx.compose.runtime.*
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.awt.awtEventOrNull
@@ -22,9 +19,9 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.input.pointer.*
 import androidx.compose.ui.layout.*
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import viewModel.CanvasViewModel
-import viewModel.UndirectedViewModel
+import view.canvas.CanvasView
+import viewModel.canvas.CanvasViewModel
+import viewModel.graph.UndirectedViewModel
 
 val HEADER_HEIGHT = 50f
 val MENU_WIDTH = 100f
