+++ src/main/kotlin/view/canvas/CanvasView.kt
@@ -1,9 +1,9 @@
-package view
+package view.canvas
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import viewModel.CanvasViewModel
+import viewModel.canvas.CanvasViewModel
 
 @Composable
 fun <V> CanvasView(
