+++ src/main/kotlin/view/canvas/VertexCanvasView.kt
@@ -1,4 +1,4 @@
-package view
+package view.canvas
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -8,7 +8,7 @@ import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import viewModel.VertexCanvasViewModel
+import viewModel.canvas.VertexCanvasViewModel
 
 @Composable
 fun <V> VertexCanvasView(
