+++ src/main/kotlin/view/graph/EdgeView.kt
@@ -1,4 +1,4 @@
-package view
+package view.graph
 
 import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.fillMaxSize
@@ -7,22 +7,21 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import viewModel.EdgeCanvasViewModel
+import viewModel.graph.EdgeViewModel
 
 @Composable
-fun <V> EdgeCanvasView(
-    viewModel: EdgeCanvasViewModel<V>,
-    modifier: Modifier = Modifier
+fun <V> EdgeView(
+    viewModel: EdgeViewModel<V>
 ) {
     Canvas(Modifier.fillMaxSize()) {
         drawLine(
             start = Offset(
-                viewModel.first.offset.x.dp.toPx() + viewModel.first.radius.toPx(),
-                viewModel.first.offset.y.dp.toPx() + viewModel.first.radius.toPx(),
+                viewModel.first.x.dp.toPx() + viewModel.first.radius.toPx(),
+                viewModel.first.y.dp.toPx() + viewModel.first.radius.toPx(),
             ),
             end = Offset(
-                viewModel.second.offset.x.dp.toPx() + viewModel.second.radius.toPx(),
-                viewModel.second.offset.y.dp.toPx() + viewModel.second.radius.toPx(),
+                viewModel.second.x.dp.toPx() + viewModel.second.radius.toPx(),
+                viewModel.second.y.dp.toPx() + viewModel.second.radius.toPx(),
             ),
             color = Color.Black,
             strokeWidth = 1f.dp.toPx()
