+++ src/main/kotlin/view/graph/UndirectedGraphView.kt
@@ -1,10 +1,10 @@
-package view
+package view.graph
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import viewModel.UndirectedViewModel
+import viewModel.graph.UndirectedViewModel
 
 @Composable
 fun <V> UndirectedGraphView(
