+++ src/main/kotlin/view/graph/VertexView.kt
@@ -1,4 +1,4 @@
-package view
+package view.graph
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
@@ -10,7 +10,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import viewModel.VertexViewModel
+import viewModel.graph.VertexViewModel
 
 @Composable
 fun <V> VertexView(
