+++ src/main/kotlin/viewModel/canvas/EdgeCanvasViewModel.kt
@@ -1,4 +1,4 @@
-package viewModel
+package viewModel.canvas
 
 class EdgeCanvasViewModel<V>(
     val first: VertexCanvasViewModel<V>,
